Implement volume progress for manga
