+++ app/components/users/library/library-list/library-entry.js
@@ -1,7 +1,6 @@
 import Component from 'ember-component';
 import get from 'ember-metal/get';
 import set from 'ember-metal/set';
-import { scheduleOnce } from 'ember-runloop';
 import createChangeset from 'ember-changeset-cp-validations';
 import { task, timeout } from 'ember-concurrency';
 import { invokeAction } from 'ember-invoke-action';
@@ -13,11 +12,6 @@ export default Component.extend({
   init() {
     this._super(...arguments);
     this.changeset = createChangeset(get(this, 'libraryEntry'));
-  },
-
-  willDestroyElement() {
-    this._super(...arguments);
-    this._removeClickHandler();
   },
 
   saveTask: task(function* (useTimeout = false) {
@@ -34,20 +28,8 @@ export default Component.extend({
   }).restartable(),
 
   actions: {
-    sanitizeNumber(value) {
-      const parsed = parseInt(value, 10);
-      return isNaN(parsed) ? value : parsed;
-    },
-
-    showInput() {
-      set(this, 'showProgressEditor', true);
-      scheduleOnce('afterRender', () => {
-        this._setupClickHandler();
-      });
-    },
-
-    incrementProgress() {
-      this.incrementProperty('changeset.progress', 1);
+    incrementProgress(attribute = 'progress') {
+      this.incrementProperty(`changeset.${attribute}`, 1);
       get(this, 'saveTask').perform(true);
     },
 
@@ -59,31 +41,6 @@ export default Component.extend({
     checkedEntry(value) {
       const libraryEntry = get(this, 'libraryEntry');
       invokeAction(this, 'checkedEntry', libraryEntry, value);
-    }
-  },
-
-  /**
-   * Setup a click event handler on the body to remove the progress input and save the
-   * changeset.
-   *
-   * @private
-   */
-  _setupClickHandler() {
-    this.progressInputHandler = ({ target }) => {
-      const isProgressInput = target.matches('.library-progress-input *, .library-progress-input');
-      if (!isProgressInput) {
-        set(this, 'showProgressEditor', false);
-        this._removeClickHandler();
-        get(this, 'saveTask').perform();
-      }
-    };
-    document.body.addEventListener('click', this.progressInputHandler);
-  },
-
-  _removeClickHandler() {
-    if (this.progressInputHandler) {
-      document.body.removeEventListener('click', this.progressInputHandler);
-      this.progressInputHandler = null;
     }
   }
 });
