+++ app/templates/components/users/library/library-list/library-entry.hbs
@@ -59,37 +59,29 @@
 
   {{! Progress }}
   <td class="progress-cell">
-    <span>
-      {{#if showProgressEditor}}
-        {{ui/selectable-input changeset.progress
-          min=0
-          max=libraryEntry.media.unitCount
-          update=(pipe-action
-            (action "sanitizeNumber")
-            (action (mut changeset.progress))
-          )
-          class="form-control library-progress-input"
-        }}
-      {{else}}
-        <a href="#" class="link" {{action "showInput"}}>
-          {{changeset.progress}}
-        </a>
-      {{/if}}
-      /
-    </span>
-    <span>{{or libraryEntry.media.unitCount (html-safe "&mdash;")}}</span>
-
-    {{! Increment Button }}
-    {{#if (can "edit library-entry" libraryEntry)}}
-      {{#unless (eq libraryEntry.media.unitCount changeset.progress)}}
-        <a href="#" class="increment-episode" {{action "incrementProgress"}}>
-          <svg xmlns="http://www.w3.org/2000/svg" width="26" height="26" viewBox="0 0 26 26">
-            <path d="M13.5,3.188C7.805,3.188,3.188,7.805,3.188,13.5S7.805,23.813,13.5,23.813S23.813,19.195,23.813,13.5 S19.195,3.188,13.5,3.188z M19,15h-4v4h-3v-4H8v-3h4V8h3v4h4V15z"></path>
-          </svg>
-        </a>
-      {{/unless}}
-    {{/if}}
+    {{users/library/library-list/progress-cell
+      progress=changeset.progress
+      max=libraryEntry.media.unitCount
+      libraryEntry=libraryEntry
+      update=(action (mut changeset.progress))
+      increment=(action "incrementProgress")
+      save=(perform saveTask)
+    }}
   </td>
+
+  {{! Volumes (Manga) }}
+  {{#if (eq libraryEntry.media.modelType "manga")}}
+    <td class="progress-cell">
+      {{users/library/library-list/progress-cell
+        progress=changeset.volumesOwned
+        max=libraryEntry.media.volumeCount
+        libraryEntry=libraryEntry
+        update=(action (mut changeset.volumesOwned))
+        increment=(action "incrementProgress" "volumesOwned")
+        save=(perform saveTask)
+      }}
+    </td>
+  {{/if}}
 
   {{! Edit }}
   {{#if (can "edit library-entry" libraryEntry)}}
