+++ app/templates/components/users/library/library-list/progress-cell.hbs
@@ -1,0 +1,28 @@
+<span>
+  {{#if showProgressEditor}}
+    {{ui/selectable-input progress
+      min=0
+      max=max
+      update=(pipe-action
+        (action "sanitizeNumber")
+        (action update)
+      )
+      class="form-control library-progress-input"
+    }}
+  {{else}}
+    <a href="#" class="link" {{action "showInput"}}>{{progress}}</a>
+  {{/if}}
+  /
+</span>
+<span>{{or max (html-safe "&mdash;")}}</span>
+
+{{! Increment Button }}
+{{#if (can "edit library-entry" libraryEntry)}}
+  {{#unless (eq libraryEntry.status "completed")}}
+    <a href="#" class="increment-episode" {{action increment}}>
+      <svg xmlns="http://www.w3.org/2000/svg" width="26" height="26" viewBox="0 0 26 26">
+        <path d="M13.5,3.188C7.805,3.188,3.188,7.805,3.188,13.5S7.805,23.813,13.5,23.813S23.813,19.195,23.813,13.5 S19.195,3.188,13.5,3.188z M19,15h-4v4h-3v-4H8v-3h4V8h3v4h4V15z"></path>
+      </svg>
+    </a>
+  {{/unless}}
+{{/if}}
