+++ translations/en-us.yaml
@@ -413,6 +413,7 @@ users:
           other {}}"
         rating: "Rating"
         progress: "Progress"
+        volumes: "Volumes"
     edit:
       status: "Library Status"
       progress: "Progress"
